Move hpimportance config into config package
